refactor(generator): float weights with running-mean anchor
Three coupled changes to sharpen intra-layer balancing.

  * ClusterData.weight, NodeEntry.weight, weight_delta are now float.
    generate_clusters keeps one decimal of precision (round(..., 1))
    instead of int rounding, matching zone_metadata.toml's float input.
  * pick_cluster_weight_matched widens its tolerance in 0.5 steps and
    accepts a float max_tolerance. The matcher uses an epsilon on the
    candidate comparison so float anchors (which arise from running
    means) don't fall through a half-step boundary.
  * pick_layer_clusters anchors each slot on the running mean of all
    prior picks (including type fallbacks) instead of pinning to
    picks[0].weight. Returns weight_deltas alongside picks/fallbacks so
    the caller doesn't reconstruct anchors.

StructureConfig.max_weight_tolerance is now float and is actually
threaded through to the matcher; it must be a non-negative multiple
of 0.5 (matcher grid). Display rounds weights/deltas to 2 decimals
to hide floating-point artifacts.

Measured: on 20 fresh seeds, 0 picks fall outside the default
max_tolerance=3.0 (vs. ~6 with the previous int+first-pick anchor),
and the dw distribution is much finer (dw=0.1, 0.2, 0.43, etc.).

+++ speedfog/clusters.py
@@ -50,7 +50,7 @@ class ClusterData:
     id: str
     zones: list[str]
     type: str  # start, final_boss, legacy_dungeon, mini_dungeon, boss_arena
-    weight: int
+    weight: float
     entry_fogs: list[dict]  # [{"fog_id": str, "zone": str}, ...]
     exit_fogs: list[dict]  # [{"fog_id": str, "zone": str, "unique"?: bool}, ...]
     unique_exit_fogs: list[dict] = field(
@@ -86,7 +86,7 @@ class ClusterData:
             id=data["id"],
             zones=data["zones"],
             type=data["type"],
-            weight=data["weight"],
+            weight=float(data["weight"]),
             entry_fogs=entry_fogs,
             exit_fogs=exit_fogs,
             unique_exit_fogs=[f for f in all_exits if f.get("unique")],
